+++ spec/support/schema.rb
@@ -4,7 +4,7 @@ ActiveRecord::Schema.define do
   create_table "double_entry_account_balances", :force => true do |t|
     t.string     "account", :null => false
     t.string     "scope"
-    t.integer    "balance", :null => false
+    t.bigint     "balance", :null => false
     t.timestamps            :null => false
   end
 
@@ -15,8 +15,8 @@ ActiveRecord::Schema.define do
     t.string     "account",         :null => false
     t.string     "scope"
     t.string     "code",            :null => false
-    t.integer    "amount",          :null => false
-    t.integer    "balance",         :null => false
+    t.bigint     "amount",          :null => false
+    t.bigint     "balance",         :null => false
     t.references "partner",                         :index => false
     t.string     "partner_account", :null => false
     t.string     "partner_scope"
@@ -40,7 +40,7 @@ ActiveRecord::Schema.define do
     t.integer    "week"
     t.integer    "day"
     t.integer    "hour"
-    t.integer    "amount",                        :null => false
+    t.bigint     "amount",                        :null => false
     t.string     "filter"
     t.string     "range_type",      :limit => 15, :null => false
     t.timestamps                                  :null => false
